requirements.txt
torch>=1.12.0
torchaudio>=0.12.0
transformers>=4.20.0
openai-whisper>=20230918
datasets>=2.0.0
librosa>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scikit-learn>=1.1.0
tensorboard>=2.8.0
jupyter>=1.0.0
tqdm>=4.64.0
pyyaml>=6.0
soundfile>=0.10.0
<<<<<<< HEAD
jiwer>=2.3.0
torch-audio>=0.12.0
accelerate>=0.20.0
evaluate>=0.4.0jiwer
matplotlib  
seaborn
datasets
datasets
=======
openai-whisper>=20230918
>>>>>>> 0e8924f53d63892523407296246ab8b27da79c95
streamlit
plotly
